# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Clerkship API - ClinicAI UNAB
  description: '**API REST de la plataforma clínica Clerkship / ClinicAI UNAB** para
    el entrenamiento y evaluación

    del proceso de razonamiento diagnóstico en estudiantes de medicina y ciencias
    de la salud.


    El backend combina:

    * **PostgreSQL** para la persistencia relacional transaccional (usuarios, cursos,
    artículos, carpetas y evaluaciones).

    * **MongoDB** para almacenamiento documental semiestructurado (transcripción de
    chats clínicos, notas de voz y documentos en base64).

    * **JWT (JSON Web Tokens)** con control de acceso basado en roles (`STUDENT`,
    `TEACHER`, `ADMIN`).

    '
  version: 1.0.0
  contact:
    name: Equipo de Desarrollo Clerkship
    url: https://github.com/Steven08Ar/Clerkship
  license:
    name: GNU General Public License v3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
- url: http://localhost:5000
  description: Servidor local de desarrollo Flask
...
paths:
...
  /api/documentos/carpetas:
    get:
      tags:
      - Documentos y Carpetas
      summary: Listar carpetas de documentos
      description: Retorna las carpetas creadas por el usuario con conteo de archivos.
      operationId: listFolders
      security:
      - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentFolderResponse'
          description: Carpetas del usuario obtenidas con éxito
    post:
      tags:
      - Documentos y Carpetas
      summary: Crear carpeta
      operationId: createFolder
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFolderRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentFolderResponse'
          description: Carpeta creada exitosamente
  /api/documentos/carpetas/{id}:
    patch:
      tags:
      - Documentos y Carpetas
      summary: Editar carpeta (nombre o color)
      operationId: updateFolder
      security:
      - BearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Casos Rotación 2026
                color:
                  type: string
                  example: '#3B82F6'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentFolderResponse'
          description: Carpeta actualizada exitosamente
    delete:
      tags:
      - Documentos y Carpetas
      summary: Eliminar carpeta
      operationId: deleteFolder
      security:
      - BearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Carpeta eliminada
  /api/documentos/archivos:
    get:
      tags:
      - Documentos y Carpetas
      summary: Listar archivos
      description: Filtra por folder_id o muestra los documentos recientes del usuario.
      operationId: listDocuments
      security:
      - BearerAuth: []
      parameters:
      - name: folder_id
        in: query
        schema:
          type: string
          format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentFileResponse'
          description: Documentos obtenidos con éxito
    post:
      tags:
      - Documentos y Carpetas
      summary: Subir documento
      description: Almacena metadatos en PostgreSQL y contenido base64 en MongoDB.
      operationId: uploadDocument
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadDocumentRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentFileResponse'
          description: Documento subido y procesado exitosamente
  /api/documentos/archivos/{id}:
    get:
      tags:
      - Documentos y Carpetas
      summary: Descargar o ver archivo
      operationId: getDocumentById
      security:
      - BearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentFileResponse'
          description: Contenido y metadatos del documento obtenidos con éxito
    delete:
      tags:
      - Documentos y Carpetas
      summary: Eliminar archivo
      operationId: deleteDocument
      security:
      - BearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Archivo eliminado
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Token de acceso JWT obtenido al iniciar sesión (`/api/auth/login`)
        o registrarse.

        Debe enviarse en la cabecera: `Authorization: Bearer <token>`.

        '
...
  schemas:
...
    CreateFolderRequest:
      type: object
      required:
      - name
      properties:
        name:
          type: string
          example: Rotación Hospitalaria
        color:
          type: string
          example: '#10B981'
    DocumentFolderResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Rotación Hospitalaria
        color:
          type: string
          example: '#10B981'
        files_count:
          type: integer
          example: 5
    UploadDocumentRequest:
      type: object
      required:
      - name
      - file_base64
      properties:
        folder_id:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
          example: Caso_01_Pancreatitis.pdf
        mime_type:
          type: string
          example: application/pdf
        file_base64:
          type: string
          description: Contenido codificado en Base64
    DocumentFileResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        folder_id:
          type: string
          nullable: true
        size_bytes:
          type: integer
        mime_type:
          type: string
        created_at:
          type: string
          format: date-time
...